requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.5.0
streamlit
